Revert "ci: test architecture integer type sizes with new RNG"
This reverts commit 13191ff8d9732102921cc998bf7e9acab6b3bbf6.

+++ src/common.jl
@@ -30,7 +30,7 @@ const ROOT_3_OVER_3 = 0.577350269189626
 ### Random number generation
 
 struct RandomState
-    seed::Int
+    seed::UInt64
     rng::Xoroshiro128Star
 end
